The htree.xls conflict was resolved by accepting the incoming version.  I am working on that file and I will handle the conflict when I merge this update back into my working copy.

The branch is closed.

See https://sourceforge.net/apps/trac/bigdata/ticket/294

git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4523 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-jini/src/java/com/bigdata/service/jini/util/DumpFederation.java
@@ -1321,13 +1321,13 @@ public class DumpFederation {
          * The #of journals that are currently under management for the data
          * service on which the index partition resides.
          */
-        public final int managedJournalCount;
+        public final long managedJournalCount;
 
         /**
          * The #of index segments that are currently under management for the
          * data service on which the index partition resides.
          */
-        public final int managedSegmentCount;
+        public final long managedSegmentCount;
 
         /**
          * The #of overflow events.
